chore: guest opt
