+++ Cargo.toml
@@ -6,7 +6,7 @@ members = [
   "prover-scheduler",
   "metrics",
   "prover-core",
-  "guests/revm",
+  "guests/guest-revm",
   "guests/aggregation",
 ]
 resolver = "2"
