+++ guests/guest-revm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "evm"
+name = "guest-revm"
 version = "0.1.0"
 edition = "2021"
 
@@ -10,3 +10,4 @@ serde = { version = "1.0", default-features = false, features = ["alloc", "deriv
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 k256 = { version = "0.13.3", features = ["ecdsa"], default-features = false }
 sp1-zkvm = { git = "https://github.com/succinctlabs/sp1.git", tag = "v4.1.3" }
+serde_cbor = "0.11.2"
