+++ prover/Cargo.toml
@@ -16,6 +16,9 @@ recursion-gnark-ffi   = { git = "https://github.com/0xEigenLabs/eigen-zkvm", bra
 
 once_cell = "1.18"
 
+models = { git = "https://github.com/0xEigenLabs/powdr-revme", branch = "continuations", package = "models" }
+serde_cbor = "0.11.2"
+
 #sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", tag = "v4.1.3", features = ["native-gnark"]}
 sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", tag = "v4.1.3" }
 bincode = "1.3.3"
@@ -42,6 +45,10 @@ tokio = { version = "1.21.0", features = ["macros", "rt-multi-thread", "signal",
 uuid = { version = "1.2", features = ["v4", "fast-rng", "macro-diagnostics"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+
+#serde = { version = "1.0", default-features = false, features = ["alloc", "derive", "rc"] }
+#serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
+
 anyhow = "1.0"
 log = "0.4.0"
