+++ prover/build.rs
@@ -1,4 +1,4 @@
 fn main() {
-    sp1_build::build_program("../guests/revm");
+    sp1_build::build_program("../guests/guest-revm");
     sp1_build::build_program("../guests/aggregation");
 }
