+++ prover/suite.json
@@ -1,0 +1,85 @@
+{
+    "_info": null,
+    "chain_id": 1,
+    "env": {
+        "currentCoinbase": "0x8943545177806ed17b9f23f0a21ee5948ecaa776",
+        "currentDifficulty": "0x0",
+        "currentGasLimit": "0x17d7840",
+        "currentNumber": "0x2",
+        "currentTimestamp": "0x682ebdaf",
+        "currentBaseFee": "0x2da282a8",
+        "previousHash": "0x5b98f1e1dc3d760508d175e0167733992e65242e0444c475c13e4281f3505637",
+        "currentRandom": "0x0000000000000000000000000000000000000000000000000000000000000000",
+        "currentBeaconRoot": "0x0000000000000000000000000000000000000000000000000000000000000000",
+        "currentWithdrawalsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+        "parentBlobGasUsed": "0x5208",
+        "parentExcessBlobGas": "0x5208"
+    },
+    "pre": {
+        "0x8943545177806ed17b9f23f0a21ee5948ecaa776": {
+            "balance": "0x33b2e3c9fd0803ce8000000",
+            "code": "0x",
+            "nonce": "0x0",
+            "storage": {}
+        },
+        "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266": {
+            "balance": "0x0",
+            "code": "0x",
+            "nonce": "0x0",
+            "storage": {}
+        }
+    },
+    "post": {
+        "Shanghai": [
+            {
+                "expectException": null,
+                "indexes": {
+                    "data": 0,
+                    "gas": 0,
+                    "value": 0
+                },
+                "hash": "0x8abc4bcc647ec3b371ec90f4a42e9140adb5f6f1c261f72adfa0b614d60530ef",
+                "postState": {
+                    "0x8943545177806ed17b9f23f0a21ee5948ecaa776": {
+                        "balance": "0x33b2e37340913700a021ac0",
+                        "code": "0x",
+                        "nonce": "0x1",
+                        "storage": {}
+                    },
+                    "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266": {
+                        "balance": "0x56bc75e2d63100000",
+                        "code": "0x00",
+                        "nonce": "0x0",
+                        "storage": {}
+                    }
+                },
+                "logs": "0x0000000000000000000000000000000000000000000000000000000000000000",
+                "txbytes": "0x7b2263616c6c6572223a22307838393433353435313737383036656431376239663233663061323165653539343865636161373736222c226761735f6c696d6974223a32313030302c226761735f7072696365223a2230783264613238326139222c227472616e736163745f746f223a7b2243616c6c223a22307866333966643665353161616438386636663463653661623838323732373963666666623932323636227d2c2276616c7565223a2230783536626337356532643633313030303030222c2264617461223a223078222c226e6f6e6365223a302c22636861696e5f6964223a312c226163636573735f6c697374223a5b5d2c226761735f7072696f726974795f666565223a22307831222c22626c6f625f686173686573223a5b5d2c226d61785f6665655f7065725f626c6f625f676173223a6e756c6c7d"
+            }
+        ]
+    },
+    "transaction": {
+        "data": [
+            "0x"
+        ],
+        "gasLimit": [
+            "0x5208"
+        ],
+        "gasPrice": "0x2da282a9",
+        "nonce": "0x0",
+        "secretKey": "0x0000000000000000000000000000000000000000000000000000000000000000",
+        "sender": "0x8943545177806ed17b9f23f0a21ee5948ecaa776",
+        "to": "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+        "value": [
+            "0x56bc75e2d63100000"
+        ],
+        "maxFeePerGas": "0x77359401",
+        "maxPriorityFeePerGas": "0x1",
+        "accessLists": [
+            []
+        ],
+        "blobVersionedHashes": [],
+        "maxFeePerBlobGas": null
+    },
+    "out": null
+}
