+++ prover/suite_opt.json
@@ -1,0 +1,87 @@
+{
+    "0x5e31d41c8394ea45c1b9dd3e47137716838c0078fdafb4a5ae74a22cbbccbedf": {
+        "_info": null,
+        "chain_id": 1,
+        "env": {
+            "currentCoinbase": "0x8943545177806ed17b9f23f0a21ee5948ecaa776",
+            "currentDifficulty": "0x0",
+            "currentGasLimit": "0x17d7840",
+            "currentNumber": "0x2",
+            "currentTimestamp": "0x682ebdaf",
+            "currentBaseFee": "0x2da282a8",
+            "previousHash": "0x5b98f1e1dc3d760508d175e0167733992e65242e0444c475c13e4281f3505637",
+            "currentRandom": "0x0000000000000000000000000000000000000000000000000000000000000000",
+            "currentBeaconRoot": "0x0000000000000000000000000000000000000000000000000000000000000000",
+            "currentWithdrawalsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+            "parentBlobGasUsed": "0x5208",
+            "parentExcessBlobGas": "0x5208"
+        },
+        "pre": {
+            "0x8943545177806ed17b9f23f0a21ee5948ecaa776": {
+                "balance": "0x33b2e3c9fd0803ce8000000",
+                "code": "0x",
+                "nonce": "0x0",
+                "storage": {}
+            },
+            "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266": {
+                "balance": "0x0",
+                "code": "0x",
+                "nonce": "0x0",
+                "storage": {}
+            }
+        },
+        "post": {
+            "Shanghai": [
+                {
+                    "expectException": null,
+                    "indexes": {
+                        "data": 0,
+                        "gas": 0,
+                        "value": 0
+                    },
+                    "hash": "0x8abc4bcc647ec3b371ec90f4a42e9140adb5f6f1c261f72adfa0b614d60530ef",
+                    "postState": {
+                        "0x8943545177806ed17b9f23f0a21ee5948ecaa776": {
+                            "balance": "0x33b2e37340913700a021ac0",
+                            "code": "0x",
+                            "nonce": "0x1",
+                            "storage": {}
+                        },
+                        "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266": {
+                            "balance": "0x56bc75e2d63100000",
+                            "code": "0x00",
+                            "nonce": "0x0",
+                            "storage": {}
+                        }
+                    },
+                    "logs": "0x0000000000000000000000000000000000000000000000000000000000000000",
+                    "txbytes": "0xac6663616c6c6572548943545177806ed17b9f23f0a21ee5948ecaa776696761735f6c696d6974195208696761735f70726963655820000000000000000000000000000000000000000000000000000000002da282a96b7472616e736163745f746fa16443616c6c54f39fd6e51aad88f6f4ce6ab8827279cfffb922666576616c756558200000000000000000000000000000000000000000000000056bc75e2d63100000646461746140656e6f6e63650068636861696e5f6964016b6163636573735f6c69737480706761735f7072696f726974795f666565582000000000000000000000000000000000000000000000000000000000000000016b626c6f625f68617368657380746d61785f6665655f7065725f626c6f625f676173f6"
+                }
+            ]
+        },
+        "transaction": {
+            "data": [
+                "0x"
+            ],
+            "gasLimit": [
+                "0x5208"
+            ],
+            "gasPrice": "0x2da282a9",
+            "nonce": "0x0",
+            "secretKey": "0x0000000000000000000000000000000000000000000000000000000000000000",
+            "sender": "0x8943545177806ed17b9f23f0a21ee5948ecaa776",
+            "to": "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266",
+            "value": [
+                "0x56bc75e2d63100000"
+            ],
+            "maxFeePerGas": "0x77359401",
+            "maxPriorityFeePerGas": "0x1",
+            "accessLists": [
+                []
+            ],
+            "blobVersionedHashes": [],
+            "maxFeePerBlobGas": null
+        },
+        "out": null
+    }
+}
